CustomLinkedList перенесен в отдельный класс
